requirements.txt
cwltool>=3.1.20231207132642
cwltest>=2.6.20250314152537
future>=0.16.0
requests>=2.18.2
PyJWT>=1.6.4
typing_extensions>=3.7.4
minio>=7.2.0
oscar_python>=1.3.1
pydantic>=2.0.0
python-dotenv>=0.19.0
